#40 VS Compilability #38 Slice Map Projection
- compiles fine on VS now, just some warnings
- WIP on Slice Map Projection fragment shader, has to use GBuffer, so
should be added to other executables then

+++ src/executables/VoxelizationCompute/VoxelGridTools.cpp
@@ -19,7 +19,7 @@ VoxelGridGPU::~VoxelGridGPU() {
 }
 
 void VoxelGridGPU::setUniformCellSizeFromResolutionAndMapping(float width,
-		float height, float resX, float resY, float resZ) {
+		float height, int resX, int resY, int resZ) {
 
 	if ( ( width / resX )  != ( height / resY) )
 	{
